Refactor: extract shared compact kernel machinery to common header

+++ src/cqr_potrf_compact.hpp
@@ -11,7 +11,7 @@
  * potrf_compact_general() -- column- or row-major, lower or upper.
  *
  * A portable, vectorized mkl_?potrf_compact, reusing the pack<T,V> / vsqrt /
- * BatchView machinery from cqr_compact.hpp (shared with geqrf/ormqr). Paired with
+ * BatchView machinery from cqr_compact_common.hpp (shared with geqrf/ormqr). Paired with
  * MKL's mkl_?trsm_compact it factors and solves batched SPD systems.
  *
  * Algorithm: the unblocked LAPACK potf2, right-looking, V matrices at a time.
@@ -50,7 +50,7 @@
 #ifndef CQR_POTRF_COMPACT_HPP
 #define CQR_POTRF_COMPACT_HPP
 
-#include "cqr_compact.hpp" /* pack<T,V>, BatchView, make_view, vsqrt<T,V> */
+#include "cqr_compact_common.hpp" /* pack<T,V>, BatchView, make_view, vsqrt<T,V> */
 
 #include <cstddef>
 #include <cassert>
@@ -231,8 +231,7 @@ void potrf_compact_general(bool rowmajor, bool upper, Int n, T *ap, Int ldap, In
         else
             /* strided: sweep the reflectorless potf2 with row stride ldap,
              * column stride 1 -- column-major upper and row-major lower. */
-            potrf_compact_group_strided<T, V, Int>(
-                n, make_view<T, V, Int>(a, static_cast<std::size_t>(ldap), 1));
+            potrf_compact_group_strided<T, V, Int>(n, make_view<T, V, Int>(a, ldap, 1));
     }
 }
